requirements.txt
astropy>=5
numpy>=1.20
PyYAML>=6
